Time: 2 ms (93.74%), Space: 43.3 MB (84.08%) - LeetHub

+++ 4-median-of-two-sorted-arrays/4-median-of-two-sorted-arrays.java
@@ -26,7 +26,7 @@ class Solution {
             k++;
         }
         if((m+n)%2!=0) {
-            return a[(m+n)/2];
+            return (double)a[(m+n)/2];
         }
         else {
             int x = (m+n)/2;
